Made default handling of invocations synchronous.
